separate out dispatcher from sc-core into separate model
