+++ src/event-stream-k8/Cargo.toml
@@ -1,0 +1,27 @@
+[package]
+name = "event-stream-k8"
+edition = "2018"
+version = "0.1.0"
+authors = ["Fluvio Contributors <team@fluvio.io>"]
+description = "Fluvio Event Stream Stroe access"
+repository = "https://github.com/infinyon/fluvio"
+license = "Apache-2.0"
+
+[dependencies]
+log = "0.4.8"
+serde = { version ="1.0.103", features = ['derive'] }
+futures = { version = "0.3.1" }
+async-trait = "0.1.21"
+async-lock = "1.1.2"
+async-channel = "1.1.0"
+event-listener = "2.2.0"
+tokio = { version = "0.2.21", features = ["macros"] }
+flv-util = { version = "0.2.0"}
+flv-future-aio = { version = "2.3.0", features =["unstable"] }
+k8-metadata-client = { version = "1.0.1"}
+flv-types = { path ="../types", version = "0.2.0"}
+flv-eventstream-model = { features = ["k8"], path = "../event-stream-model" }
+
+[dev-dependencies]
+flv-future-aio = { version = "2.3.0", features=["fixture"]}
+flv-util = { version = "0.2.0", features=["fixture"]}
