+++ src/event-stream-k8/src/lib.rs
@@ -1,0 +1,17 @@
+pub mod store;
+pub mod dispatcher;
+pub mod actions;
+
+mod error;
+
+pub use error::StoreError;
+
+pub mod core {
+    pub use flv_eventstream_model::core::*;
+}
+
+
+pub mod k8 {
+
+    pub use flv_eventstream_model::k8::*;
+}
