+++ src/event-stream-model/Cargo.toml
@@ -3,7 +3,7 @@ name = "flv-eventstream-model"
 edition = "2018"
 version = "0.1.0"
 authors = ["Fluvio Contributors <team@fluvio.io>"]
-description = "Fluvio metadata"
+description = "Fluvio Event Stream Model"
 repository = "https://github.com/infinyon/fluvio"
 license = "Apache-2.0"
 
@@ -18,7 +18,6 @@ k8-obj-metadata = { version = "1.0.0", optional = true }
 k8-obj-core = { version = "1.1.0", optional = true}
 k8-obj-app = { version = "1.1.0", optional = true }
 flv-future-aio = { version = "2.3.0"}
-kf-protocol =  "2.0.0" 
 flv-types = { path ="../types", version = "0.2.0" }
 flv-util = { version = "0.2.0"}
