+++ src/metadata-cluster/src/lib.rs
@@ -17,15 +17,6 @@ pub mod store {
 #[cfg(feature = "k8")]
 pub mod k8 {
 
-    pub mod core {
-        pub use k8_obj_core::*;
-    }
-
-    pub mod app {
-        pub use k8_obj_app::*;
-    }
-
-    pub mod metadata {
-        pub use k8_obj_metadata::*;
-    }
+    pub use flv_eventstream_model::k8::*;
+    
 }
