+++ src/metadata-cluster/src/topic/mod.rs
@@ -31,4 +31,19 @@ mod metadata {
     impl Creatable for TopicSpec {}
 
     impl Status for TopicStatus {}
+
+    #[cfg(feature = "k8")]
+    mod extended {
+    
+        use super::TopicSpec;
+        use crate::store::k8::K8ExtendedSpec;
+
+        impl K8ExtendedSpec for TopicSpec {
+            type K8Spec = Self;
+            type K8Status = Self::Status;
+        }
+
+        
+    }
+
 }
