+++ src/sc-core/Cargo.toml
@@ -30,6 +30,7 @@ kf-service = { path = "../kf-service"}
 internal-api = { path = "../internal-api"}
 flv-metadata-cluster = { features = ["k8"],path = "../metadata-cluster"}
 sc-api = { version = "2.0.0", path = "../sc-api",  package = "flv-api-sc"}
+event-stream-k8 = { path = "../event-stream-k8" }
 
 [dev-dependencies]
 flv-future-aio = { version = "2.3.0", features=["fixture"]}
