+++ src/sc-core/src/stores/partition.rs
@@ -1,12 +1,6 @@
 pub use flv_metadata_cluster::partition::store::*;
 pub use flv_metadata_cluster::partition::*;
-
-use super::*;
+pub use flv_metadata_cluster::store::k8::K8MetaItem;
 
 pub type PartitionAdminMd = PartitionMetadata<K8MetaItem>;
 pub type PartitionAdminStore = PartitionLocalStore<K8MetaItem>;
-
-impl K8ExtendedSpec for PartitionSpec {
-    type K8Spec = Self;
-    type K8Status = Self::Status;
-}
