+++ src/sc-core/src/stores/spu.rs
@@ -1,15 +1,11 @@
 pub use flv_metadata_cluster::spu::store::*;
 pub use flv_metadata_cluster::spu::*;
+pub use flv_metadata_cluster::store::k8::K8MetaItem;
 
-use super::*;
 
 pub type SpuAdminMd = SpuMetadata<K8MetaItem>;
 pub type SpuAdminStore = SpuLocalStore<K8MetaItem>;
 
-impl K8ExtendedSpec for SpuSpec {
-    type K8Spec = Self;
-    type K8Status = Self::Status;
-}
 
 // check if given range is conflict with any of the range
 pub async fn is_conflict(
