+++ src/sc-core/src/stores/topic.rs
@@ -1,12 +1,7 @@
 pub use flv_metadata_cluster::topic::store::*;
 pub use flv_metadata_cluster::topic::*;
-
-use super::*;
+pub use flv_metadata_cluster::store::k8::K8MetaItem;
 
 pub type TopicAdminStore = TopicLocalStore<K8MetaItem>;
 pub type TopicAdminMd = TopicMetadata<K8MetaItem>;
 
-impl K8ExtendedSpec for TopicSpec {
-    type K8Spec = Self;
-    type K8Status = Self::Status;
-}
